refactor : seperate quotes and quotes slider component
